+++ CHANGELOG.md
@@ -2,6 +2,11 @@ Important changes to data models, configuration, and migrations between each
 AppEngine version, listed here to ease deployment and troubleshooting.
 
 ## Next Release (replace with git tag when deployed)
+ * Bumped runtimeVersion to `2023.01.16`
+ * Upgraded stable Dart analysis SDK to `2.18.7`.
+ * Upgraded preview Dart analysis SDK to `2.19.0-444.4.beta`.
+ * Upgraded preview Flutter analysis SDK to `3.7.0-1.4.pre`.
+ * Upgraded dependencies.
 
 ## `20230112t112900-all`
  * Bumped runtimeVersion to `2023.01.11`